Moved isDateInterval and interfaces to index and helper
