+++ packages/ui-lib/src/helpers/index.tsx
@@ -1,5 +1,6 @@
 import { isEqual } from 'date-fns';
-import {DateInterval, isDateInterval, ITimeUnit} from '../index';
+import { ITimeUnit } from '..';
+import { DateInterval } from '../Filters';
 
 const isInsideRange = (value: number, min: number, max: number) : boolean => {
   if( value < min) {return false;}
@@ -111,9 +112,27 @@ const uniqueID = () =>
     return false;
   };
 
+
+  const isDateInterval = (value: any): value is DateInterval => {
+    if (value === null || value === undefined) {
+      return false;
+    }
+  
+    if (value.start === null || value.start === undefined) {
+      return false;
+    }
+  
+    if (value.end === null || value.end === undefined) {
+      return false;
+    }
+  
+    return (value.start instanceof Date) && (value.end instanceof Date);
+  };
+
 export {
   clamp,
   isValidImage,
+  isDateInterval,
   getImageType,
   getTextTimeUnit,
   isInsideRange,
